wip: NFT refactoring - splitting to modules
